+++ mobile/src/main/res/values/strings.xml
@@ -61,6 +61,25 @@
     <string name="connected_state_valid">%s (reachable)</string>
     <string name="connected_state_failed">%s (lost)</string>
 
+    <string name="clients_popup_nickname">Nickname…</string>
+    <string name="clients_popup_block">Block</string>
+    <string name="clients_popup_unblock">Unblock</string>
+    <string name="clients_popup_stats">Stats…</string>
+    <string name="clients_nickname_title">Nickname for %s</string>
+    <string name="clients_stats_title">Stats for %s</string>
+    <plurals name="clients_stats_message_1">
+        <item quantity="one">Connected 1 time since %2$s</item>
+        <item quantity="other">Connected %1$s times since %2$s</item>
+    </plurals>
+    <plurals name="clients_stats_message_2">
+        <item quantity="one">Sent 1 packet, %2$s</item>
+        <item quantity="other">Sent %1$s packets, %2$s</item>
+    </plurals>
+    <plurals name="clients_stats_message_3">
+        <item quantity="one">Received 1 packet, %2$s</item>
+        <item quantity="other">Received %1$s packets, %2$s</item>
+    </plurals>
+
     <string name="settings_upstream">Upstream</string>
     <string name="settings_downstream">Downstream</string>
     <string name="settings_service_masquerade">IP Masquerade</string>
